requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
astroid==2.11.6
asyncpg==0.25.0
attrs==21.4.0
black==22.3.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.5.18.1
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
conda==4.3.16
crashtest==0.3.1
cryptography==3.4.8
dill==0.3.5.1
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
filelock==3.7.1
flake8==4.0.1
h11==0.13.0
html5lib==1.1
httptools==0.4.0
identify==2.5.1
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
isort==5.10.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==21.8.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
MarkupSafe==2.1.1
mccabe==0.6.1
msgpack==1.0.3
mypy-extensions==0.4.3
nodeenv==1.6.0
orjson==3.7.2
packaging==20.9
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
pipenv==2022.5.2
pkginfo==1.8.2
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.1.13
poetry-core==1.0.8
pre-commit==2.19.0
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
pylev==1.4.0
pylint==2.14.3
pyparsing==3.0.9
pypika-tortoise==0.1.5
pytest==7.1.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tortoise-orm==0.19.1
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
watchgod==0.8.2
webencodings==0.5.1
websockets==10.3
wrapt==1.14.1
